requirements.txt
flask==2.3.3
flask-wtf==1.1.1
flask-session==0.8.0
PyPDF2==3.0.1
google-generativeai==0.3.1
reportlab==4.0.4
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.32.3
gunicorn==21.2.0
